moved arena pages out of components
