+++ src/App.tsx
@@ -19,12 +19,12 @@ import AccessDenied from "./containers/AccessDeniedPage/AccessDeniedPage";
 import AllSubjectsPage from "./containers/AllSubjectsPage/AllSubjectsPage";
 import ErrorPage from "./containers/ErrorPage/ErrorPage";
 import ArenaAdminPage from "./containers/MyNdla/Arena/ArenaAdminPage";
+import ArenaFlagPage from "./containers/MyNdla/Arena/ArenaFlagPage";
 import ArenaNotificationPage from "./containers/MyNdla/Arena/ArenaNotificationsPage";
 import ArenaPage from "./containers/MyNdla/Arena/ArenaPage";
+import ArenaSingleFlagPage from "./containers/MyNdla/Arena/ArenaSingleFlagPage";
+import ArenaUserListPage from "./containers/MyNdla/Arena/ArenaUserListPage";
 import CategoryEditPage from "./containers/MyNdla/Arena/CategoryEditPage";
-import ArenaFlagPage from "./containers/MyNdla/Arena/components/ArenaFlagPage";
-import ArenaSingleFlagPage from "./containers/MyNdla/Arena/components/ArenaSingleFlagPage";
-import ArenaUserListPage from "./containers/MyNdla/Arena/components/ArenaUserListPage";
 import NewCategoryPage from "./containers/MyNdla/Arena/NewCategoryPage";
 import { NewTopicPage } from "./containers/MyNdla/Arena/NewTopicPage";
 import PostsPage from "./containers/MyNdla/Arena/PostsPage";
